pdf-generator-backend/requirements.txt
Flask==3.0.3
Pillow==10.2.0
reportlab==4.2.2
Flask-Cors==5.0.0
